+++ EventListener/BugsnagListener.php
@@ -5,14 +5,14 @@ namespace Bugsnag\BugsnagBundle\EventListener;
 use Bugsnag\BugsnagBundle\Request\SymfonyResolver;
 use Bugsnag\Client;
 use Bugsnag\Report;
+use Symfony\Component\Console\ConsoleEvents;
+use Symfony\Component\Console\Event\ConsoleErrorEvent;
 use Symfony\Component\Console\Event\ConsoleExceptionEvent;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\Event\GetResponseForExceptionEvent;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
-
 use Symfony\Component\HttpKernel\KernelEvents;
-use Symfony\Component\Console\ConsoleEvents;
-use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 class BugsnagListener implements EventSubscriberInterface
 {
@@ -80,25 +80,7 @@ class BugsnagListener implements EventSubscriberInterface
      */
     public function onKernelException(GetResponseForExceptionEvent $event)
     {
-        if (!$this->auto) {
-            return;
-        }
-
-        $exception = $event->getException();
-
-        $report = Report::fromPHPThrowable(
-            $this->client->getConfig(),
-            $exception
-        );
-        $report->setUnhandled(true);
-        $report->setSeverityReason([
-            'type' => 'unhandledExceptionMiddleware',
-            'attributes' => [
-                'framework' => 'Symfony',
-            ],
-        ]);
-
-        $this->client->notify($report);
+        $this->sendNotify($event->getException(), []);
     }
 
     /**
@@ -110,22 +92,38 @@ class BugsnagListener implements EventSubscriberInterface
      */
     public function onConsoleException(ConsoleExceptionEvent $event)
     {
+        $meta = [ 'status' => $event->getExitCode() ];
+        if ($event->getCommand()) {
+            $meta['name'] = $event->getCommand()->getName();
+        }
+        $this->sendNotify($event->getException(), ['command' => $meta]);
+    }
+
+    /**
+     * Handle a console error.
+     *
+     * @param \Symfony\Component\Console\Event\ConsoleErrorEvent $event
+     *
+     * @return void
+     */
+    public function onConsoleError(ConsoleErrorEvent $event)
+    {
+        $meta = [ 'status' => $event->getExitCode() ];
+        if ($event->getCommand()) {
+            $meta['name'] = $event->getCommand()->getName();
+        }
+        $this->sendNotify($event->getError(), ['command' => $meta]);
+    }
+
+    private function sendNotify($throwable, $meta)
+    {
         if (!$this->auto) {
             return;
         }
 
-        $exception = $event->getException();
-
-        $meta = [
-            'command' => [
-                'name' => $event->getCommand()->getName(),
-                'status' => $event->getExitCode(),
-            ],
-        ];
-
         $report = Report::fromPHPThrowable(
             $this->client->getConfig(),
-            $exception
+            $throwable
         );
         $report->setUnhandled(true);
         $report->setSeverityReason([
@@ -141,12 +139,18 @@ class BugsnagListener implements EventSubscriberInterface
 
     public static function getSubscribedEvents(){
         $listeners = [
-            KernelEvents::REQUEST => ['onKernelRequest',256],
-            KernelEvents::EXCEPTION => ['onException',128]
+            KernelEvents::REQUEST => ['onKernelRequest', 256],
+            KernelEvents::EXCEPTION => ['onKernelException', 128]
         ];
 
+        // Added ConsoleEvents in Symfony 2.3
         if (class_exists('Symfony\Component\Console\ConsoleEvents')) {
-            $listeners[class_exists('Symfony\Component\Console\Event\ConsoleErrorEvent') ? ConsoleEvents::ERROR : ConsoleEvents::EXCEPTION] = ['onConsoleException',128];
+            // Added with ConsoleEvents::ERROR in Symfony 3.3 to deprecate ConsoleEvents::EXCEPTION
+            if (class_exists('Symfony\Component\Console\Event\ConsoleErrorEvent')) {
+                $listeners[ConsoleEvents::ERROR] = ['onConsoleError', 128];
+            } else {
+                $listeners[ConsoleEvents::EXCEPTION] = ['onConsoleException', 128];
+            }
         }
 
         return $listeners;