+++ Tests/Listener/BugsnagListenerTest.php
@@ -9,7 +9,13 @@ use Bugsnag\Report;
 use GrahamCampbell\TestBenchCore\MockeryTrait;
 use Mockery;
 use PHPUnit_Framework_TestCase as TestCase;
+
+use Exception;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Event\ConsoleErrorEvent;
 use Symfony\Component\Console\Event\ConsoleExceptionEvent;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\HttpKernel\Event\GetResponseForExceptionEvent;
 
 class ReportStub
@@ -38,11 +44,48 @@ class BugsnagListenerTest extends TestCase
         $report->shouldReceive('setUnhandled')->once()->with(true);
         $report->shouldReceive('setSeverityReason')->once()->with(['type' => 'unhandledExceptionMiddleware', 'attributes' => ['framework' => 'Symfony']]);
         $client->shouldReceive('getConfig')->once()->andReturn('config');
+        $report->shouldReceive('setMetaData')->once()->with([]);
         $client->shouldReceive('notify')->once()->with($report);
 
         // Initiate test
         $listener = new BugsnagListener($client, $resolver, true);
         $listener->onKernelException($event);
+    }
+
+    /**
+     * @requires class foo
+     */
+    public function testOnConsoleError()
+    {
+        if (!class_exists('Symfony\Component\Console\Event\ConsoleErrorEvent')) {
+            $this->markTestSkipped('ConsoleErrorEvent class not present - pre-Symfony3.3');
+        } else {
+            // Create mocks
+            $report = Mockery::namedMock(Report::class, ReportStub::class);
+            $client = Mockery::mock(Client::class);
+            $input = Mockery::mock(InputInterface::class);
+            $output = Mockery::mock(OutputInterface::class);
+            $command = Mockery::mock(Command::class);
+            $event = new ConsoleErrorEvent($input, $output, new Exception(), $command); // Unable to mock as final
+            $event->setExitCode(1);
+            $resolver = Mockery::mock(SymfonyResolver::class);
+
+            // Setup responses
+            $command->shouldReceive('getName')->once()->andReturn('test');
+            $report->shouldReceive('setMetaData')->once()->with(['command' => ['name' => 'test', 'status' => 1]]);
+            $report->shouldReceive('fromPHPThrowable')
+                ->with('config', 'exception')
+                ->once()
+                ->andReturn($report);
+            $report->shouldReceive('setUnhandled')->once()->with(true);
+            $report->shouldReceive('setSeverityReason')->once()->with(['type' => 'unhandledExceptionMiddleware', 'attributes' => ['framework' => 'Symfony']]);
+            $client->shouldReceive('getConfig')->once()->andReturn('config');
+            $client->shouldReceive('notify')->once()->with($report);
+
+            // Initiate test
+            $listener = new BugsnagListener($client, $resolver, true);
+            $listener->onConsoleError($event);
+        }
     }
 
     public function testOnConsoleException()
@@ -55,7 +98,7 @@ class BugsnagListenerTest extends TestCase
 
         // Setup responses
         $event->shouldReceive('getException')->once()->andReturn('exception');
-        $event->shouldReceive('getCommand')->once()->andReturn($event);
+        $event->shouldReceive('getCommand')->twice()->andReturn($event);
         $event->shouldReceive('getName')->once()->andReturn('test');
         $event->shouldReceive('getExitCode')->once()->andReturn(1);